backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
google-generativeai==0.8.3
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==2.1.3
faiss-cpu==1.9.0.post1
reportlab==4.2.5
matplotlib==3.9.4
httpx==0.28.1
python-multipart==0.0.17
pytest==8.3.4
